MWC_huggingface/requirements.txt
# Core Dependencies
#torch>=2.0.0
transformers>=4.38.0
datasets>=2.14.0
evaluate>=0.4.0
accelerate>=0.24.0

# LoRA and Quantization
peft>=0.6.0
bitsandbytes>=0.41.0

# Audio Processing
soundfile>=0.12.1
librosa>=0.10.0

# Evaluation Metrics
sacrebleu>=2.3.1
jiwer>=3.0.0

# Utilities
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.26.0,<2.0
pyarrow>=12.0.0
tqdm>=4.65.0

# HuggingFace Hub
huggingface_hub>=0.20.0
